+++ src/1_gathering/db/tppt/downloading.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 # -*- coding: utf-8 -*-
 
+if [ ! -f ../data/external/dbSource/tppt/TPPT_database.xlsx ]; then
+
 mkdir -p ../data/external/dbSource/tppt/
 
 wget "https://www.agroscope.admin.ch/dam/agroscope/de/dokumente/publikationen/tppt-xls.xlsx.download.xlsx/TPPT_database.xlsx" -O ../data/external/dbSource/tppt/TPPT_database.xlsx
+
+fi
+
+echo "Done"
